shim, primitives: move namespace validation logic
The logic is moved from the pallet blobs into primitives
to share it with the shim and provide useful error messaging
when an invalid namespace is provided in the submit command
